Add tests to cover Npgsql v6
