+++ ElasticApmAgent.sln
@@ -181,6 +181,10 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RabbitMqSample", "sample\Ra
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OpenTelemetrySample", "sample\OpenTelemetrySample\OpenTelemetrySample.csproj", "{3D31589A-CA44-4D6D-9944-26CE9640A283}"
 EndProject
+Project("{D954291E-2A0B-460D-934E-DC6B0785DB48}") = "Npgsql6Common", "sample\Npgsql6Common\Npgsql6Common.shproj", "{42D8402D-19B0-4EC5-A347-919157DD29A4}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Npgsql6Sample", "sample\Npgsql6Sample\Npgsql6Sample.csproj", "{58D8B98E-C52E-47F4-8DF2-3CBED2BDC490}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -471,6 +475,10 @@ Global
 		{3D31589A-CA44-4D6D-9944-26CE9640A283}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3D31589A-CA44-4D6D-9944-26CE9640A283}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{3D31589A-CA44-4D6D-9944-26CE9640A283}.Release|Any CPU.Build.0 = Release|Any CPU
+		{58D8B98E-C52E-47F4-8DF2-3CBED2BDC490}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{58D8B98E-C52E-47F4-8DF2-3CBED2BDC490}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{58D8B98E-C52E-47F4-8DF2-3CBED2BDC490}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{58D8B98E-C52E-47F4-8DF2-3CBED2BDC490}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -547,6 +555,8 @@ Global
 		{2B23487A-B340-4F5C-A49B-9B829F437A5A} = {3C791D9C-6F19-4F46-B367-2EC0F818762D}
 		{1D6B0C67-42C8-4AB4-A795-B6FF8EF7196E} = {3C791D9C-6F19-4F46-B367-2EC0F818762D}
 		{3D31589A-CA44-4D6D-9944-26CE9640A283} = {3C791D9C-6F19-4F46-B367-2EC0F818762D}
+		{42D8402D-19B0-4EC5-A347-919157DD29A4} = {3C791D9C-6F19-4F46-B367-2EC0F818762D}
+		{58D8B98E-C52E-47F4-8DF2-3CBED2BDC490} = {3C791D9C-6F19-4F46-B367-2EC0F818762D}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {69E02FD9-C9DE-412C-AB6B-5B8BECC6BFA5}
